feat: refactor enqueueMessage to it's own inngest function
