+++ src/app/api/inngest/route.ts
@@ -30,6 +30,7 @@ import {
   unstopConfirmationSMSCommunicationJourney,
   welcomeSMSCommunicationJourney,
 } from '@/inngest/functions/sms'
+import { enqueueSMS } from '@/inngest/functions/sms/enqueueMessages'
 import { deleteUserActions } from '@/inngest/functions/user/deleteUserActions'
 import {
   auditUsersTotalDonationAmountUsdInngest,
@@ -75,5 +76,6 @@ export const { GET, POST, PUT } = serve({
     backfillReactivationCron,
     sendEventNotificationWithInngest,
     deleteUserActions,
+    enqueueSMS,
   ],
 })
